Moved code to controller
