+++ backend/api/controllers/account.js
@@ -2,9 +2,11 @@ const mongoose = require('mongoose');
 const Account = require('../models/account');
 const bcrpyt = require('bcrypt');
 const jwt = require('jsonwebtoken');
+const UserID = require('../models/userid');
+
 
 exports.signup = (req, res, next) => {
-    Account.find({ email: req.body.email })
+    Account.find({ email: req.query.email })
         .exec()
         .then(account => {
             if (account.length >= 1) {
@@ -12,20 +14,22 @@ exports.signup = (req, res, next) => {
                     message: 'An account with this email already exist'
                 });
             } else {
-                bcrpyt.hash(req.body.password, 10, (err, hash) => {
+                bcrpyt.hash(req.query.password, 10, (err, hash) => {
                     if (err) {
                         return res.status(500).json({
                             error: err
                         });
                     } else {
+                        newid().then(id => {
                         const account = new Account({
                             _id: new mongoose.Types.ObjectId(),
-                            email: req.body.email,
+                            userid : id,
+                            email: req.query.email,
                             password: hash,
-                            username: req.body.username,
-                            label: req.body.label,
-                            bio: req.body.bio,
-                            profilePic: req.body.profilePic,
+                            username: req.query.username,
+                            label: req.query.label,
+                            bio: req.query.bio,
+                            profilePic: req.query.profilePic,
                             verified: false
                         })
                         account.save()
@@ -52,6 +56,7 @@ exports.signup = (req, res, next) => {
                                 error: err
                             })
                         });
+                    });
                     }
                 });
             }
@@ -65,7 +70,7 @@ exports.signup = (req, res, next) => {
 }
 
 exports.login = (req, res, next) => {
-    Account.find({ email: req.body.email })
+    Account.find({ email: req.query.email })
     .exec()
     .then(account => {
         if (account.length < 1) {
@@ -73,7 +78,7 @@ exports.login = (req, res, next) => {
                 message: "Email not found, user does not exist"
             });
         }
-        bcrpyt.compare(req.body.password, account[0].password, (err, result) => {
+        bcrpyt.compare(req.query.password, account[0].password, (err, result) => {
             if (err) {
                 return res.status(401).json({
                     message: "Auth failed"
@@ -252,3 +257,18 @@ exports.delete = (req, res, next) => {
         });
     });
 }
+
+//Generate new ID for user
+const newid = async () =>{
+    const currentVal = await UserID.getCurrent();
+    const newID = currentVal+1;
+    UserID.updateOne({name : "Counter"}, {current : newID})
+    .exec()
+    .then(err => {
+        if (err) {
+            console.error(err);
+          } 
+    });
+    return newID;
+}
+
