+++ backend/api/routes/memory.js
@@ -65,8 +65,8 @@ router.get('/:memoryID', (req, res, next) => {
 });
 
 router.patch('/like', checkAuth, (req, res, next) => {
-    const memoryID = req.body.memoryID;
-    const accountID = req.body.accountID;
+    const memoryID = req.query.memoryID;
+    const accountID = req.query.accountID;
     const memory = Memory.findById(memoryID);
     if (!memory) {
         return res.status(404).json({
